DEP-2763 : Separate out the proactive controller test
